Implement new GUC swap_limit
To make configuration easier, we diverge from the kernel's choice
to represent the limit as "memory + swap" which is not only confusing,
but forces you to change the values in the correct order so that
they always match.

This makes things easier for the user, but harder for us:
we have to do the whole dance of getting the current values,
adjusting them correctly and setting them in the right order.

This desperately asks for regression tests, which should be
added soon.