Don't try to read kernel parameters
GOCs are under control of PostgreSQL, so changing them on the fly
is not a good idea.  This is particularly true for strings, which
have to be freed and allocated by guc.c.

Hence, remove all attempts to modify GUCs while they are displayed.

This makes the module simpler, but removes some of its power.

Document also the ALTER SYSTEM ... RESET cannot change the kernel,
because no hooks are called in that case.