Get rid of the dependency on "libcgroup"
Relying on libcgroup has several disadvantages:
- it adds an extra requirement for building pg_cgroups
- the library's interface is clumsy
- the resulting code is clumsy and inefficient
- RedHat has deprecated the library

This patch gets rid of the dependency and implements control group
support by directly interfacing with the kernel via the special cgroup
file systems.

The code that deals with this is placed into a new file "libcg1.c"
(I chose this name in anticipation of a move to cgroups v2 at some
later time).

The resulting code is only aroung 150 lines more in total, and some
of that is boilerplate because of the split into several source files,
so I feel justified on that end alone.  Good riddance, libcgroup!

Also fix a bug that used 100000 as default value and maximum for
"pg_cgroups.cpu_share".  100000 corresponds to one CPU core, and since
there can be more than one of those on a machine, the limit is really
a multiple of that.  Unfortunately that forces me to reduce the regression
test coverage, because defaut value and maximum are now machine dependent.