Call back release 1.0.0 and fix bugs
1.0.0 was pretty premature, the software is not ready for prime time:
because of a bug, the control group value weren't actually changed

The regression tests don't check that, hence the problem, but I don't
know how to query the operating system from the regression tests.

Also, add a new (read-only) parameter "pg_cgroups.version".

Bug left for later: deletion of the control group is not working.